revert main back to 7.5.1
ran too fast with too few tests

+++ src/prng.h
@@ -26,6 +26,6 @@ int prng(unsigned int max);
 
 uint64_t prngGetSeed(void);
 
-uint64_t prngSetSeed(uint64_t seed);
+int prngSetSeed(uint64_t seed);
 
 #endif //FAR_HORIZONS_PRNG_H
